badge

+++ ponos/ponos/blas/c_regular_grid.inl
@@ -22,16 +22,16 @@
  *
  */
 
-template <typename T>
+template<typename T>
 CRegularGrid<T>::CRegularGrid(const size3 &d, const T &b, const vec3 cellSize,
                               const vec3 &offset) {
   this->dimensions = d;
   this->background = b;
   data = new T **[d[0]];
-  for (int i = 0; i < d[0]; i++)
+  for (int i = 0; i < static_cast<i64>(d[0]); i++)
     data[i] = new T *[d[1]];
-  for (int i = 0; i < d[0]; i++)
-    for (int j = 0; j < d[1]; j++)
+  for (int i = 0; i < static_cast<i64>(d[0]); i++)
+    for (int j = 0; j < static_cast<i64>(d[1]); j++)
       data[i][j] = new T[d[2]];
   this->toWorld.reset();
   this->toWorld.scale(cellSize.x, cellSize.y, cellSize.z);
@@ -40,7 +40,7 @@ CRegularGrid<T>::CRegularGrid(const size3 &d, const T &b, const vec3 cellSize,
   this->toGrid = inverse(this->toWorld);
 }
 
-template <typename T>
+template<typename T>
 CRegularGrid<T>::CRegularGrid(const size3 &d, const T &b, const bbox3 &bb) {
   this->dimensions = d;
   this->background = b;
@@ -61,7 +61,7 @@ CRegularGrid<T>::CRegularGrid(const size3 &d, const T &b, const bbox3 &bb) {
   this->toGrid = inverse(this->toWorld);
 }
 
-template <typename T> CRegularGrid<T>::~CRegularGrid() {
+template<typename T> CRegularGrid<T>::~CRegularGrid() {
   for (u32 i = 0; i < this->dimensions[0]; i++)
     for (u32 j = 0; j < this->dimensions[1]; j++)
       delete[] data[i][j];
@@ -70,24 +70,24 @@ template <typename T> CRegularGrid<T>::~CRegularGrid() {
   delete[] data;
 }
 
-template <typename T> void CRegularGrid<T>::setAll(T v) {
+template<typename T> void CRegularGrid<T>::setAll(T v) {
   for (auto ijk : Index3Range<i32>(this->dimensions))
     data[ijk[0]][ijk[1]][ijk[2]] = v;
 }
 
-template <typename T> T CRegularGrid<T>::safeData(int i, int j, int k) const {
+template<typename T> T CRegularGrid<T>::safeData(int i, int j, int k) const {
   return data[max(0, min(this->dimensions[0] - 1, i))]
-             [max(0, min(this->dimensions[1] - 1, j))]
-             [max(0, min(this->dimensions[2] - 1, k))];
+  [max(0, min(this->dimensions[1] - 1, j))]
+  [max(0, min(this->dimensions[2] - 1, k))];
 }
 
-template <typename T> void CRegularGrid<T>::set(const size3 &i, const T &v) {
+template<typename T> void CRegularGrid<T>::set(const size3 &i, const T &v) {
   this->data[std::max(0u, std::min(this->dimensions[0] - 1, i[0]))]
-            [std::max(0u, std::min(this->dimensions[1] - 1, i[1]))]
-            [std::max(0u, std::min(this->dimensions[2] - 1, i[2]))] = v;
+  [std::max(0u, std::min(this->dimensions[1] - 1, i[1]))]
+  [std::max(0u, std::min(this->dimensions[2] - 1, i[2]))] = v;
 }
 
-template <typename T> void CRegularGrid<T>::normalize() {
+template<typename T> void CRegularGrid<T>::normalize() {
   T M = data[0][0][0];
   for (auto ijk : Index3Range<i32>(this->dimensions))
     M = ponos::max(M, data[ijk[0]][ijk[1]][ijk[2]]);
@@ -95,20 +95,20 @@ template <typename T> void CRegularGrid<T>::normalize() {
     data[ijk[0]][ijk[1]][ijk[2]] /= M;
 }
 
-template <typename T> void CRegularGrid<T>::normalizeElements() {
+template<typename T> void CRegularGrid<T>::normalizeElements() {
   for (auto ijk : Index3Range<i32>(this->dimensions))
     data[ijk[0]][ijk[1]][ijk[2]] =
         ponos::normalize(data[ijk[0]][ijk[1]][ijk[2]]);
 }
 
-template <typename T> CRegularGrid2D<T>::CRegularGrid2D() {}
+template<typename T> CRegularGrid2D<T>::CRegularGrid2D() {}
 
-template <typename T>
+template<typename T>
 CRegularGrid2D<T>::CRegularGrid2D(uint32_t w, uint32_t h) {
   this->setDimensions(w, h);
 }
 
-template <typename T> T CRegularGrid2D<T>::sample(float x, float y) const {
+template<typename T> T CRegularGrid2D<T>::sample(float x, float y) const {
   point2 gp = this->toGrid(point2(x, y));
   int x0 = static_cast<int>(gp.x);
   int y0 = static_cast<int>(gp.y);
@@ -127,7 +127,7 @@ template <typename T> T CRegularGrid2D<T>::sample(float x, float y) const {
   return ponos::bicubicInterpolate<T>(p, gp.x - x0, gp.y - y0);
 }
 
-template <typename T> void CRegularGrid2D<T>::updateDataStructure() {
+template<typename T> void CRegularGrid2D<T>::updateDataStructure() {
   data.resize(this->width, std::vector<T>());
   for (size_t i = 0; i < this->width; i++)
     data[i].resize(this->height);
